+++ src/users_groups_lib/entities/group.py
@@ -1,8 +1,6 @@
 """This entity represents a unix group."""
 from dataclasses import dataclass
 
-from users_groups_lib.entities import User
-
 
 @dataclass
 class Group:
@@ -11,8 +9,8 @@ class Group:
     Attributes:
         gid: The identification of group.
         name: The name of the group.
-        users: List of users in group.
+        users: List of users names in group.
     """
     gid: int
     name: str
-    users: list[User]
+    users: list[str]