+++ src/users_groups_lib/managers/getters/__init__.py
@@ -1,2 +1,3 @@
 """Exposed getters classes and methods."""
 from users_groups_lib.managers.getters.user_getter import UserGetter
+from users_groups_lib.managers.getters.group_getter import GroupGetter