+++ src/users_groups_lib/managers/group_manager.py
@@ -2,20 +2,22 @@
 
 from shell_executor_lib import CommandManager
 
+from users_groups_lib import Group
+from users_groups_lib.managers.getters import GroupGetter
+
 
 class GroupManager:
     """Manage unix groups in shell."""
 
-    def __init__(self, user: str, password: str) -> None:
+    def __init__(self, command_manager: CommandManager) -> None:
         """Initialize the GroupManager.
 
         Args:
-            user: The user of the shell who execute command.
-            password: The password of the user.
+            command_manager: To make commands in the shell.
         """
-        self._command_manager = CommandManager(user, password)
+        self.__group_getter: GroupGetter = GroupGetter(command_manager)
 
-    async def get_groups(self) -> None:
+    async def get_groups(self) -> list[Group]:
         """Obtain the groups from the shell in a list.
 
         Returns:
@@ -24,15 +26,4 @@ class GroupManager:
         Raises:
             CommandError: If the exit code is not 0.
         """
-        pass
-
-    async def get_group(self) -> None:
-        """Obtain the groups from the shell in a list.
-
-        Returns:
-            A list of the groups in the shell.
-
-        Raises:
-            CommandError: If the exit code is not 0.
-        """
-        pass
+        return await self.__group_getter.get_groups()