+++ tests/mock_users_groups_lib/__init__.py
@@ -1,4 +1,7 @@
 """Exposed mock_users_groups_lib classes and methods."""
-from mock_users_groups_lib.mocks_user_manager import (mock_users_list, mock_group_name_list,
-                                                      mock_command_executor_method, mock_user, mock_user_entity)
-from mock_users_groups_lib.mocks_group_manager import mock_group
+from mock_users_groups_lib.mocks_user_manager import (mock_users_list, mock_group_name_list, mock_user,
+                                                      mock_user_entity, mock_users_list_entities_without_group_name,
+                                                      mock_user_entity_group_name, mock_users_list_entities)
+from mock_users_groups_lib.mocks_group_manager import mock_group, mock_groups_entities, mock_groups_entity
+
+mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"