+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - @javierfh03 - Develop add user in user manager with unit tests.
 - @javierfh03 - Develop edit user in user manager with unit tests.
 - @javierfh03 - Develop delete user in user manager with unit tests.
+- @javierfh03 - Develop get groups in group manager with unit tests and little changes in user sections.
 
 ### Changed
 - @javierfh03 - Change structure of library, missing any unit test.